fix: testcases 75-82 for ctrip query flights
Because the tests only differ in the from/to and date, they encounter
the same issues. The main issue is not selecting one-way. It is
important to put the one-way button selection action just before
clicking search because the agent might accidentally select and set the
'return date' field, and clicking the one-way button automatically
clears that field too. Another issue is that the agent sometimes fills
the WRONG from field for FROM/TO. So in tests 76,77,78, I explicitly described the
location of the fields to set. For unknown reasons, this modification
doesn't always work, so it is not applied to other tests. Finally, I
made the final 'expected result' more flexible.